cleaned up gradle files

+++ core/src/com/monstergoboom/game/interfaces/services/RenderService.kt
@@ -2,5 +2,5 @@ package com.monstergoboom.game.interfaces.services
 
 interface RenderService {
     fun initialize();
-    fun update(delta: Long): Long;
+    fun update(delta: Float): Long;
 }
